// apps/frontend/src/types/payload-types.ts
export type HeadScripts =
  | {
      type: "external" | "inline" | "custom";
      src?: string | null;
      content?: string | null;
      async?: boolean | null;
      defer?: boolean | null;
      id?: string | null;
    }[]
  | null;
// ...
export interface User {
  id: number;
  operator?: (number | null) | Operator;
  role?: (number | null) | UserRole;
  kpi_department?: (number | null) | KpiPlansDepartment;
  restaurant?: (number | null) | Restaurant;
  access?: (number | null) | UserAccess;
  auth_method?: ("default" | "microsoft") | null;
  updatedAt: string;
  createdAt: string;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "operators".
 */
export interface Operator {
  id: number;
  name?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "user-roles".
 */
export interface UserRole {
  id: number;
  name?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "kpi-plans-department".
 */
export interface KpiPlansDepartment {
  id: number;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  name: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "restaurants".
 */
export interface Restaurant {
  id: number;
  logo?: number | Media | null;
  featured_image?: number | Media | null;
  name: string;
  slug: string;
  description?: string | null;
  description_rt?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  options?: {
    include_survey_customer?: boolean | null;
    include_employee_perks?: boolean | null;
    order?: number | null;
  };
  scripts?: HeadScripts;
  additional_data?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "user-access".
 */
export interface UserAccess {
  id: number;
  name?: string | null;
  access?:
    | {
        name?: string | null;
        hidden?: boolean | null;
        read?: boolean | null;
        create?: boolean | null;
        update?: boolean | null;
        admin?: boolean | null;
        delete?: boolean | null;
        access?: boolean | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface MenuPage {
  id: number;
  info: {
    restaurant?: (number | null) | Restaurant;
    title?: string | null;
    slug_override?: boolean | null;
    slug: string;
  };
  c: {
    font_primary?: number | MenuMedia | null;
    font_secondary?: number | MenuMedia | null;
    primary?: string | null;
    primary_contrast?: string | null;
    bg?: string | null;
    bg_card?: string | null;
    text?: string | null;
    neutral?: string | null;
    menus?: (number | Menu)[] | null;
    blk: BlockSection[];
  };
  adv: {
    className?: string | null;
    css?: string | null;
    js?: string | null;
  };
  seo: {
    title?: string | null;
    description?: string | null;
    image?: number | MenuMedia | null;
  };
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "menu-media".
 */
export interface MenuMedia {
  id: number;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "menu".
 */
export interface Menu {
  id: number;
  restaurant?: (number | null) | Restaurant;
  image?: number | MenuMedia | null;
  category_title?: string | null;
  restaurant_title?: string | null;
  slug: string;
  slug_override?: boolean | null;
  category: number | MenuCategory;
  description?: string | null;
  menu_items?:
    | {
        item: number | MenuItem;
        id?: string | null;
      }[]
    | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "menu-categories".
 */
export interface MenuCategory {
  id: number;
  title?: string | null;
  slug: string;
  slug_override?: boolean | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "menu-items".
 */
export interface MenuItem {
  id: number;
  restaurant?: (number | null) | Restaurant;
  image?: number | MenuMedia | null;
  title: string;
  slug: string;
  slug_override?: boolean | null;
  description?: string | null;
  price: number;
  availability_period?: ("breakfast" | "lunch" | "dinner" | "all_day")[] | null;
  allergen?: (number | MenuAllergen)[] | null;
  tags?: (number | MenuTag)[] | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "menu-allergens".
 */
export interface MenuAllergen {
  id: number;
  title: string;
  slug: string;
  slug_override?: boolean | null;
  description?: string | null;
  image?: number | MenuMedia | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "menu-tags".
 */
export interface MenuTag {
  id: number;
  title: string;
  slug: string;
  slug_override?: boolean | null;
  description?: string | null;
  created_by?: (number | null) | User;
  updated_by?: (number | null) | User;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockSection".
 */
export interface BlockSection {
  c: {
    blks?:
      | (BlockCarousel | BlockImages | BlockEmbed | BlockText | BlockButtons | BlockHeader | BlockFilter | BlockItems)[]
      | null;
  };
  settings: {
    className?: string | null;
    css?: string | null;
    js?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "section";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockCarousel".
 */
export interface BlockCarousel {
  c: {
    slides?:
      | {
          image?: number | MenuMedia | null;
          id?: string | null;
        }[]
      | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "carousel";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockImages".
 */
export interface BlockImages {
  c: {
    image?: number | MenuMedia | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "images";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockEmbed".
 */
export interface BlockEmbed {
  c: {
    html?: string | null;
  };
  settings: {
    css?: string | null;
    js?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "embed";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockText".
 */
export interface BlockText {
  c: {
    label?: string | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "text";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockButtons".
 */
export interface BlockButtons {
  c: {
    label?: string | null;
    link?: string | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "buttons";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockHeader".
 */
export interface BlockHeader {
  c: {
    label?: string | null;
    logo?: number | MenuMedia | null;
    show_language?: boolean | null;
    show_notification?: boolean | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "header";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockFilter".
 */
export interface BlockFilter {
  c: {
    showSearch?: boolean | null;
    showAllergenFilters?: boolean | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "filter";
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "BlockItems".
 */
export interface BlockItems {
  c: {
    title?: string | null;
    description?: string | null;
  };
  settings: {
    css?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: "items";
}